First step of reading sender email from env file
Notice that this is just a dummy email used for testing purposes and you can put yours if you want which will represents the sender email
just open env file , white down your email  on the first line without any quotes of extra space
then write down your password on the second line also without any quotes of extra space
finally make sure Not to leave a space on the third line.
just Two one for email and one for password and that's it.